+++ augur/tasks/data_analysis/__init__.py
@@ -4,3 +4,31 @@ from augur.tasks.data_analysis.discourse_analysis.tasks import discourse_analysi
 from augur.tasks.data_analysis.insight_worker.tasks import insight_model
 from augur.tasks.data_analysis.message_insights.tasks import message_insight_model
 from augur.tasks.data_analysis.pull_request_analysis_worker.tasks import pull_request_analysis_model
+
+def machine_learning_phase(logger):
+
+    with DatabaseSession(logger) as session:
+        repos = session.query(Repo).all()
+
+    ml_tasks = []
+    clustering_tasks = []
+    discourse_tasks = []
+    insight_tasks = []
+    message_insights_tasks = []
+    pull_request_analysis_tasks = []
+    for repo in repos:
+        clustering_tasks.append(clustering_model.si(repo.repo_git))
+        discourse_tasks.append(discourse_analysis_model.si(repo.repo_git))
+        insight_tasks.append(insight_model.si(repo.repo_git))
+        message_insights_tasks.append(message_insight_model.si(repo.repo_git))
+        pull_request_analysis_tasks.append(pull_request_analysis_model.si(repo.repo_git))   
+
+    ml_tasks.extend(insight_tasks)
+    ml_tasks.extend(discourse_tasks)
+    ml_tasks.extend(message_insights_tasks)
+    ml_tasks.extend(pull_request_analysis_tasks)
+    ml_tasks.extend(clustering_tasks) 
+        
+    task_chain = chain(*ml_tasks)
+
+    return task_chain
