+++ augur/tasks/init/celery_app.py
@@ -2,7 +2,7 @@
 from celery.signals import worker_process_init, worker_process_shutdown
 import logging
 from typing import List, Dict
-
+import os
 from celery import Celery
 from celery import current_app 
 from celery.signals import after_setup_logger
@@ -37,7 +37,10 @@ data_analysis_tasks = ['augur.tasks.data_analysis.message_insights.tasks',
                        'augur.tasks.data_analysis.discourse_analysis.tasks',
                        'augur.tasks.data_analysis.pull_request_analysis_worker.tasks']
 
-tasks = start_tasks + github_tasks + git_tasks + data_analysis_tasks
+if os.environ.get('AUGUR_DOCKER_DEPLOY') != "1":
+    tasks = start_tasks + github_tasks + git_tasks + data_analysis_tasks
+else:
+    tasks = start_tasks + github_tasks + git_tasks
 
 redis_db_number, redis_conn_string = get_redis_conn_values()
