+++ augur/tasks/start_tasks.py
@@ -3,6 +3,7 @@ from typing import List
 import time
 import logging
 import json
+import os
 from enum import Enum
 
 from celery.result import AsyncResult
@@ -11,7 +12,8 @@ from celery import group, chain, chord, signature
 
 
 from augur.tasks.github import *
-from augur.tasks.data_analysis import *
+if os.environ.get('AUGUR_DOCKER_DEPLOY') != "1":
+    from augur.tasks.data_analysis import *
 from augur.tasks.github.detect_move.tasks import detect_github_repo_move
 from augur.tasks.github.releases.tasks import collect_releases
 from augur.tasks.github.repo_info.tasks import collect_repo_info
@@ -71,36 +73,11 @@ def repo_collect_phase(logger):
             collect_releases.si()
         )
 
-def machine_learning_phase(logger):
 
-    with DatabaseSession(logger) as session:
-        repos = session.query(Repo).all()
+DEFINED_COLLECTION_PHASES = [prelim_phase, repo_collect_phase]
+if os.environ.get('AUGUR_DOCKER_DEPLOY') != "1":
+    DEFINED_COLLECTION_PHASES.append(machine_learning_phase)
 
-    ml_tasks = []
-    clustering_tasks = []
-    discourse_tasks = []
-    insight_tasks = []
-    message_insights_tasks = []
-    pull_request_analysis_tasks = []
-    for repo in repos:
-        clustering_tasks.append(clustering_model.si(repo.repo_git))
-        discourse_tasks.append(discourse_analysis_model.si(repo.repo_git))
-        insight_tasks.append(insight_model.si(repo.repo_git))
-        message_insights_tasks.append(message_insight_model.si(repo.repo_git))
-        pull_request_analysis_tasks.append(pull_request_analysis_model.si(repo.repo_git))   
-
-    ml_tasks.extend(insight_tasks)
-    ml_tasks.extend(discourse_tasks)
-    ml_tasks.extend(message_insights_tasks)
-    ml_tasks.extend(pull_request_analysis_tasks)
-    ml_tasks.extend(clustering_tasks) 
-        
-    task_chain = chain(*ml_tasks)
-
-    return task_chain
-
-
-DEFINED_COLLECTION_PHASES = [prelim_phase, repo_collect_phase, machine_learning_phase]
 
 class AugurTaskRoutine:
     """class to keep track of various groups of collection tasks as well as how they relate to one another.
@@ -109,27 +86,19 @@ class AugurTaskRoutine:
     Attributes:
         logger (Logger): Get logger from AugurLogger
         jobs_dict (dict): Dict of data collection phases to run
-        disabled_collection_groups (List[str]): List of keys in jobs dict that have been marked as disabled
+        collection_phases (List[str]): List of phases to run in augur collection.
     """
-    def __init__(self,disabled_collection_phases: List[str]=[], disabled_collection_tasks: List[str]=[]):
+    def __init__(self,collection_phases: List[str]=[]):
         self.logger = AugurLogger("data_collection_jobs").get_logger()
         #self.session = TaskSession(self.logger)
         self.jobs_dict = {}
-        self.disabled_collection_phases = disabled_collection_phases
-        self.disabled_collection_tasks = disabled_collection_tasks
+        self.collection_phases = collection_phases
+        #self.disabled_collection_tasks = disabled_collection_tasks
 
         #Assemble default phases
         #These will then be able to be overridden through the config.
-        if prelim_phase.__name__ not in self.disabled_collection_phases:
-            self.jobs_dict[prelim_phase.__name__] = prelim_phase
-        
-        if repo_collect_phase.__name__ not in self.disabled_collection_phases:
-            self.jobs_dict[repo_collect_phase.__name__] = repo_collect_phase
-
-        if machine_learning_phase.__name__ not in self.disabled_collection_phases:
-            self.jobs_dict[machine_learning_phase.__name__] = machine_learning_phase
-
-                
+        for phase in collection_phases:
+            self.jobs_dict[phase.__name__] = phase
 
     #Get and set dict values that correspond to phases of collection
     def __getitem__(self,key: str) -> dict:
@@ -140,18 +109,8 @@ class AugurTaskRoutine:
     def __setitem__(self,key: str,newJobs):
         """Create a new collection job group with the name of the key specified.
         """
-        if key in self.disabled_collection_phases:
-            self.logger.info(f"Group {key} has been disabled")
-            return
+        self.collection_phases.append(newJobs)
         self.jobs_dict[key] = newJobs
-
-    
-    def disable_group(self,key: str):
-        """Make a group deleted from the dict and unable to be run or added.
-        """
-        del self.jobs_dict[key]
-        self.disabled_collection_phases.append(key)
-
 
     def start_data_collection(self):
         """Start all task items and return.
@@ -195,11 +154,12 @@ def start_task():
         config = session.config
         phase_options = config.get_section("Task_Routine")
 
-    #Get list of disabled phases
-    disabled_phases = [name for name, phase in phase_options.items() if phase == 0]
+    #Get list of enabled phases 
+    enabled_phase_names = [name for name, phase in phase_options.items() if phase == 1]
+    enabled_phases = [phase for phase in DEFINED_COLLECTION_PHASES if phase.__name__ in enabled_phase_names]
 
     #print(f"disabled: {disabled_phases}")
-    augur_collection = AugurTaskRoutine(disabled_collection_phases=disabled_phases)
+    augur_collection = AugurTaskRoutine(collection_phases=enabled_phases)
 
     augur_collection.start_data_collection()
